Added slashes and Moved <h2></h2>
Added slashes to tags that were supposed to be end tags, but weren't because they were missing their slashes. Also moved <h2></h2> from line 7, where it was inside the <ul> tag, for which it is not a valid child, just outside and above it so as to not change the formatting of the page.